feat: Add AI image text extraction (OCR) for file uploads
Implemented automatic text extraction for uploaded images using Google Gemini 2.0 Flash multimodal capabilities.
- Added `extractTextFromImage` to `lib/ai/claude.ts`.
- Updated `createContentAction` in `app/actions/content.ts` to process image files using refactored secure retrieval.
- Refactored file retrieval logic into `getFileBuffer` in `lib/storage.ts` to reduce duplication and centralize security checks (path traversal, IDOR).
- Added `downloadFromGCS` to `lib/storage.ts`.
- Added consolidated unit tests in `lib/ai/claude.test.ts`.

This enables users to search for text within images and allows AI auto-tagging and summarization to work on image content.
