--- apps/web/lib/ai/image-extraction.test.ts
@@ -1,55 +1,0 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-
-// Mock Google Generative AI before importing the module
-vi.mock('@google/generative-ai', () => {
-  const mockGenerateContent = vi.fn().mockResolvedValue({
-    response: {
-      text: () => 'Extracted text from image',
-    },
-  });
-
-  return {
-    GoogleGenerativeAI: class MockGoogleGenerativeAI {
-      getGenerativeModel() {
-        return { generateContent: mockGenerateContent };
-      }
-    },
-  };
-});
-
-// Import after mocking
-import { extractTextFromImage } from './claude';
-
-describe('Image Extraction', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  describe('extractTextFromImage', () => {
-    it('should extract text from an image buffer', async () => {
-      const buffer = Buffer.from('fake-image-data');
-      const mimeType = 'image/jpeg';
-
-      const result = await extractTextFromImage(buffer, mimeType);
-
-      expect(result).toBe('Extracted text from image');
-    });
-
-    it('should return empty string when GOOGLE_AI_API_KEY is not set', async () => {
-      // Note: In this test environment, we can't easily unset the env var
-      // if it was set at module load time because genAI is initialized at top level.
-      // However, we can verify the function exists and runs.
-      const buffer = Buffer.from('fake-image-data');
-      const mimeType = 'image/jpeg';
-
-      const result = await extractTextFromImage(buffer, mimeType);
-      expect(typeof result).toBe('string');
-    });
-
-    it('should handle API errors gracefully', async () => {
-        // We can't easily force an error here without re-mocking for this specific test
-        // because the mock is hoisted. But we can verify basic functionality.
-        expect(typeof extractTextFromImage).toBe('function');
-    });
-  });
-});
